+++ Solution/RiskOfChaos/EffectHandling/Controllers/ChaosEffectActivationSoundHandler.cs
@@ -1,5 +1,6 @@
 ﻿using RiskOfChaos.Content;
 using RoR2;
+using RoR2.Audio;
 using UnityEngine;
 using UnityEngine.Networking;
 
@@ -12,24 +13,24 @@ namespace RiskOfChaos.EffectHandling.Controllers
         static ChaosEffectActivationSoundHandler _instance;
         public static ChaosEffectActivationSoundHandler Instance => _instance;
 
-        const string EFFECT_ACTIVATION_SOUND_EVENT_NAME = "Play_env_hiddenLab_laptop_sequence_fail";
-
-        static bool _canPlaySound;
-        static uint _effectActivationSoundEventID;
+        static AkEventIdArg _effectActivationSoundEvent;
 
         [SystemInitializer]
         static void Init()
         {
-            _canPlaySound = !Application.isBatchMode;
+            bool canPlaySound = !Application.isBatchMode;
 
-            if (_canPlaySound)
+            AkEventIdArg effectActivationSoundEvent = 0;
+            if (canPlaySound)
             {
-                _effectActivationSoundEventID = AkSoundEngine.GetIDFromString(EFFECT_ACTIVATION_SOUND_EVENT_NAME);
-                if (_effectActivationSoundEventID == 0)
+                effectActivationSoundEvent = (AkEventIdArg)"Play_env_hiddenLab_laptop_sequence_fail";
+                if (effectActivationSoundEvent == 0)
                 {
                     Log.Error("Failed to find effect activation sound ID");
                 }
             }
+
+            _effectActivationSoundEvent = effectActivationSoundEvent;
         }
 
         ChaosEffectDispatcher _effectDispatcher;
@@ -64,30 +65,19 @@ namespace RiskOfChaos.EffectHandling.Controllers
         [ClientRpc]
         void RpcPlayEffectActivatedSound()
         {
-            PlayEffectActivatedSoundClient();
-        }
-
-        [Client]
-        public void PlayEffectActivatedSoundClient()
-        {
-            if (!_canPlaySound || _effectActivationSoundEventID == 0)
+            if (_effectActivationSoundEvent == 0)
                 return;
 
             foreach (AkAudioListener audioListener in AkAudioListener.DefaultListeners.ListenerList)
             {
-                AkSoundEngine.PostEvent(_effectActivationSoundEventID, audioListener.gameObject);
+                AkSoundEngine.PostEvent(_effectActivationSoundEvent, audioListener.gameObject);
             }
         }
-
+        
         [Server]
         public void PlayEffectActivatedSoundServer()
         {
             RpcPlayEffectActivatedSound();
-
-            if (NetworkClient.active)
-            {
-                PlayEffectActivatedSoundClient();
-            }
         }
     }
 }